В модуле validate-form.js добавила проверку типа загружаемых файлов и обработчики событий для загрузки файлов
